fix: Merge `route` and `handler` into `handler.go`
