--- src/internal/route.go
@@ -1,17 +1,0 @@
-package internal
-
-import (
-	"github.com/gin-gonic/gin"
-)
-
-type Route struct {
-	handler Handler
-}
-
-func NewRoute(handler Handler) *Route {
-	return &Route{handler}
-}
-
-func (r *Route) Route(rg *gin.RouterGroup) {
-
-}
